stdst refactor
