+++ include/skilltrees/inventory.h
@@ -5,39 +5,8 @@
 #ifndef INCLUDE_INVENTORY_H_
 #define INCLUDE_INVENTORY_H_
 
-#include "common/common.h"
-#include "skilltrees/skill.h"
+#include "skilltrees/stdst.h"
 
-/* ======================= */
-/* === DATA STRUCTURES === */
-/* ======================= */
-/* ALL the skills belonging to a player */
-typedef struct inventory {
-    // An array of active skills belonging to a player
-    skill_t** active;
-
-    // The number of active skills belonging to a player
-    unsigned int nactive;
-
-    // The maximum number of active skills a player can possess
-    // (This field helps to enforce skill tree balancing)
-    unsigned int max_active;
-
-    // An array of passive skills belonging to a player
-    skill_t** passive;
-
-    // The number of passive skills belonging to a player
-    unsigned int npassive;
-
-    // The maximum number of passive skills a player can possess
-    // (This field helps to enforce skill tree balancing)
-    unsigned int max_passive;
-
-} inventory_t;
-
-/* ================= */
-/* === FUNCTIONS === */
-/* ================= */
 /*
  * Allocates a new (empty) skill inventory in the heap.
  *
